:sparkles: Implementation of class diagram
The FooTest.java#relationsTest() is failing on local. I believe it's because same instance of Foo is used on all tests. It's making conflicts.

On CI everything is ok though. Because only classTest() and relationsTest() of the FoorTest.java file are run. In this case, there is no conflicts, so all tests pass